More Scaffolding!
Why this change is being made:
- This change adds the basic functionality for a todo list!
- It works I think

There's no routes to handle the form requests and I have some questions
around the Restfulness approach here!

As an aside, I am really enjoying gleam but this feels like way too much
work just to get like a todo app made and running. Not unproductive or
unweildy but it is probably a hallmark of not having a framework for
handling stuff like this

What were the changes made to support this:
- add a new entry to the context
- add a new page for todos
- Add a new model for TodoItem
- add some helper functions for managing todo item state