Add a new middleware for todo_items
Why this change is being made:
- This is letting us use the ctx to set a cookie.

I don't like this pattern but I want to keep this example for jwts and
potentially cors related form add ons. I think the obvious refactor is
to use a Repo pattern for setting and fetching the data (maybe hard to
do with what we have for access patterns in the cookies).

I also want to move the todo items routes to a specific handler. That
way I can keep the related code in one place and keep the main router
pretty.

What were the changes made to support this:
- Add new middleware.
- set and fetch json from cookies for our todo_items